Updated terrain generator
Now tree and rocks are spowned quite well.
Perling noise values generate randomly in a controlled range.
FIX: since the perlin noise not usually generate values as 0 or 1 probably the maps (0 to max map height) not corresponding to real values. That's probably is the problem related to the "Floating map" but idk how to solve it .... only way it's to keep track of the min and max values given during the perlin noise generation... not worth it computationally